requirements.txt
playwright==1.41.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
